fix: resolve bug in mobile view
Address an issue in mobile view by fixing the code in Aside.jsx
component. The bug caused This bug was found in the navigation menu on
the mobile view which resulted in the list menu not being displayed and
this problem was solved by adding logic so that navigation can be seen
in certain roles.

Changes:
- Modified: resources/js/Components/Sidebars/Aside.jsx